requirements.txt
openai
deepeval
ollama
sentence-transformers
chromadb
numpy
mlflow
litellm
ragas
python-dotenv

